refactor container top
refactor the top command into pkg/cmd/container
